+++ app_v2/src/main/java/com/money/manager/ex/data/repository/FakeAccountRepositoryImpl.kt
@@ -3,6 +3,7 @@ package com.money.manager.ex.data.repository
 import com.money.manager.ex.domain.model.Account
 import com.money.manager.ex.domain.model.AccountType
 import com.money.manager.ex.domain.repository.AccountRepository
+import java.math.BigDecimal
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
@@ -16,8 +17,8 @@ class FakeAccountRepositoryImpl @Inject constructor() : AccountRepository {
         
         val mockData = listOf(
 
-            Account(1, "Bank", AccountType.CHECKING, "Open", true, "€", null, 12000.0, 11500.0),
-            Account(2, "Cash", AccountType.CASH, "Open", false, "€", null, 500.0, 500.0),
+            Account(1, "Bank", AccountType.CHECKING, "Open", true, "€", null, BigDecimal.valueOf(12000.0), BigDecimal.valueOf(11500.0)),
+            Account(2, "Cash", AccountType.CASH, "Open", false, "€", null, BigDecimal.valueOf(500.0), BigDecimal.valueOf(500.0)),
             Account(
                 id = 3,
                 name = "Carta Credito",
@@ -26,8 +27,8 @@ class FakeAccountRepositoryImpl @Inject constructor() : AccountRepository {
                 isFavorite = true,
                 currencySymbol = "$",
                 currencyPrefix = null,
-                balance = -1200.0,
-                ledgerBalance = -150.0
+                balance = BigDecimal.valueOf(-1200.0),
+                ledgerBalance = BigDecimal.valueOf(-150.0)
             )
         )
         emit(mockData)
